run_process - non-block read from stdout/stderr, partial print to console
This commit modifies `run_process` subroutine.
Both stdout/stderr works in a non-blocking fashion.
When an incomplete line is received from a child process it is logged
with an extra flag 'incomplete_line'. That flag is recognized by
custom StreamHandler, which prepends "clear line" control character
and does not add termination "\n" to the printed line. That allows to
correctly propagate "progress bars"/etc to the console from child process
and still capture the output (and possibly dump it to log files).